+++ clients/admin-ui/src/features/common/table/v2/FidesTable.tsx
@@ -5,10 +5,12 @@ import {
   Td,
   Th,
   Thead,
+  Tooltip,
   Tr,
 } from "@fidesui/react";
 import {
   flexRender,
+  Row,
   RowData,
   Table as TableInstance,
 } from "@tanstack/react-table";
@@ -44,6 +46,7 @@ type Props<T> = {
   rowActionBar?: ReactNode;
   footer?: ReactNode;
   onRowClick?: (row: T) => void;
+  renderRowTooltipLabel?: (row: Row<T>) => string | undefined;
 };
 
 export const FidesTableV2 = <T,>({
@@ -51,6 +54,7 @@ export const FidesTableV2 = <T,>({
   rowActionBar,
   footer,
   onRowClick,
+  renderRowTooltipLabel,
 }: Props<T>) => (
   <TableContainer
     data-testid="fidesTable"
@@ -109,46 +113,56 @@ export const FidesTableV2 = <T,>({
       <Tbody data-testid="fidesTable-body">
         {rowActionBar}
         {tableInstance.getRowModel().rows.map((row) => (
-          <Tr
-            key={row.id}
-            height="36px"
-            _hover={
-              onRowClick
-                ? { backgroundColor: "gray.50", cursor: "pointer" }
-                : undefined
+          <Tooltip
+            key={`tooltip-${row.id}`}
+            label={
+              renderRowTooltipLabel ? renderRowTooltipLabel(row) : undefined
             }
-            data-testid={`row-${row.id}`}
+            hasArrow
+            placement="top"
           >
-            {row.getVisibleCells().map((cell) => (
-              <Td
-                key={cell.id}
-                width={
-                  cell.column.columnDef.meta?.width
-                    ? cell.column.columnDef.meta.width
-                    : "unset"
-                }
-                borderBottomWidth="1px"
-                borderBottomColor="gray.200"
-                borderRightWidth="1px"
-                borderRightColor="gray.200"
-                _first={{
-                  borderLeftWidth: "1px",
-                  borderLeftColor: "gray.200",
-                }}
-                height="inherit"
-                style={getTableTHandTDStyles(cell.column.id)}
-                onClick={
-                  cell.column.columnDef.header !== "Enable" && onRowClick
-                    ? () => {
-                        onRowClick(row.original);
-                      }
-                    : undefined
-                }
-              >
-                {flexRender(cell.column.columnDef.cell, cell.getContext())}
-              </Td>
-            ))}
-          </Tr>
+            <Tr
+              key={row.id}
+              height="36px"
+              _hover={
+                onRowClick
+                  ? { backgroundColor: "gray.50", cursor: "pointer" }
+                  : undefined
+              }
+              data-testid={`row-${row.id}`}
+              backgroundColor={row.getCanSelect() ? undefined : "gray.50"}
+            >
+              {row.getVisibleCells().map((cell) => (
+                <Td
+                  key={cell.id}
+                  width={
+                    cell.column.columnDef.meta?.width
+                      ? cell.column.columnDef.meta.width
+                      : "unset"
+                  }
+                  borderBottomWidth="1px"
+                  borderBottomColor="gray.200"
+                  borderRightWidth="1px"
+                  borderRightColor="gray.200"
+                  _first={{
+                    borderLeftWidth: "1px",
+                    borderLeftColor: "gray.200",
+                  }}
+                  height="inherit"
+                  style={getTableTHandTDStyles(cell.column.id)}
+                  onClick={
+                    cell.column.columnDef.header !== "Enable" && onRowClick
+                      ? () => {
+                          onRowClick(row.original);
+                        }
+                      : undefined
+                  }
+                >
+                  {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                </Td>
+              ))}
+            </Tr>
+          </Tooltip>
         ))}
       </Tbody>
       {footer}